HDR gain-map image output via libultrahdr, spatial video (MV-HEVC) to
mp4 conversion tooling, and a BIGINT content.filesize column.
In production for over a month.

+++ app/koken/schema.php
@@ -286,7 +286,9 @@ return  [
                     'constraint' => 255
                 ],
                 'filesize' => [
-                    'type' => 'INT'
+                    // BIGINT, not INT: a signed INT caps at 2147483647, so any
+                    // upload over 2GB fails the insert and surfaces as a 400.
+                    'type' => 'BIGINT'
                 ],
                 'width' => [
                     'type' => 'INT',
